refactor(frontend): improve type safety and reactivity in hooks and components
- Add IHasId interface and export structure for shared entities
- Update useApiQuery and useCrud hooks to support reactive query keys
- Fix card modal reactivity by updating local state on card update
- Remove unused formatDate function and handleClose method
- Change position type from int to long in UpdateCardCommand
- Remove redundant Produces attribute from CardsController

+++ backend/Organetto/Organetto.UseCases/Boards/Columns/Cards/Commands/UpdateCardCommand.cs
@@ -21,7 +21,7 @@ namespace Organetto.UseCases.Boards.Columns.Cards.Commands
         long Id,
         string? Title,
         string? Description,
-        int? Position,
+        long? Position,
         DateTimeOffset? DueDate
     ) : IRequest<CardDto>, IHasId<long>;
